Fix #27
The delivery setting is set for a pickup location. On the checkout page, the pickup location is selected but the delivery setting is loading of the global setting instead of loading of the custom delivery setting. This issue is coming when the settings for shipping methods are set first and then of pickup location.